requirements.txt
Flask
pillow
imgaug
albumentations
streamlit
numpy